+++ test/07_full_parallel.cpp
@@ -75,14 +75,14 @@ class Garage {
   NLOHMANN_DEFINE_TYPE_INTRUSIVE(Garage, cars_)
 };
 
-long double calcIncrement(const float min, const float max, const std::uint64_t timeSteps) {
+double calcIncrement(const float min, const float max, const std::uint64_t timeSteps) {
   if (min > max) {
     throw std::invalid_argument("min cannot be greater than max");
   }
   if (timeSteps == 0) {
     throw std::invalid_argument("Time steps must be greater than zero...");
   }
-  return static_cast<long double>(max - min) / static_cast<long double>(timeSteps - 1);
+  return static_cast<double>(max - min) / static_cast<double>(timeSteps - 1);
 }
 
 std::uint64_t calcDataSize(const std::uint64_t maxTimeSteps, const std::uint64_t samplingRate) {
@@ -92,8 +92,8 @@ std::uint64_t calcDataSize(const std::uint64_t maxTimeSteps, const std::uint64_t
   if (samplingRate == 0) {
     throw std::invalid_argument("Division by zero is not allowed...");
   }
-  long double result = static_cast<long double>(maxTimeSteps) /
-                       static_cast<long double>(samplingRate);
+  double result = static_cast<double>(maxTimeSteps) /
+                  static_cast<double>(samplingRate);
   return static_cast<std::uint64_t>(std::ceil(result) + 1);
 }
